Keep tenant tree entry even with empty tree
When all the objects of a Tenant are deleted in AIM, today
we also delete the entry in the TreeMapping table.
This causes the agents to completely forget about that Tenant
(thinking that some other agent was taking care of it)
leaving stale objects in ACI.

In order to avoid this, the proposal is keep the tree-mapping entry with
an empty tree value, so that the agent serving that tenant will
delete the ACI counterpart.
AID will take care of gracefully delete the DB entry when the
synchronization is complete.