Add DB model for EPG contracts
Have used an auto-increment column on the EPG table
to generate an internal ID that will be referenced
by the contracts table. This is needed for resources
that have non-scalar attributes (e.g. list of contracts).
Having an internal ID avoids having primary
keys that use many columns in the table that contains
the non-scalar values. Have updated other ACI model
classes with this internal-ID.

Signed-off-by: Amit Bose <amitbose@gmail.com>