DB based configuration for APIC/AIM options
Introduced a new CLI command: aim config [update|replace]
A config file needs to be passed with the '-c' option,
it needs to contain both DB information and the apic/aim config
sections that will populate the database.

During first time installations, 'replace' can be called with an
empty config file (except for the DB credentials) in order to
set all the default values.