Avoid deleting hash subtree during update
When updating a node in the hash-tree,
AimHashTreeMaker deletes the subtree at that node
to remove child nodes that are directly
related (e.g. relation-nodes). But this ends
up deleting other unrelated child nodes as
well (those that correspond to other AIM
resources). This problem is fixed by adding
a metadata flag to all child-nodes that are
related to a node, and it is only these
related nodes that are removed during an
update.

Signed-off-by: Amit Bose <amitbose@gmail.com>