Retrieve children from events
When events are received from the WS interface on a particular
APIC object, we need to retrieve all the interesting related
items in order to properly build the hash tree.
More specifically, when a RS gets modified from an APIC object,
the latter should be completely retrieved together with all its
references.