options without default values should be commented
Options with example values should be commented. Orchestration tools only change values that are required. So the service fails to start, if the orchestration has not changed a value. Example. if rabbit_port=<rabbit-m1-port> is not commented, and orchestration does not change it, the service will fail to start