Reconnect Web-Socket on ACI error 405
Occasionally, APIC lose track of an open web-socket and is unable
to open subscriptions for any new URLs, returning a 405 error.
In these cases, we should reconnect the web-socket for all the
tenants in AID.

If this fix appears to be insufficient to recover from such
cases, we should consider killing AID after a certain number
of failures.

Closes issue noironetworks/support #499