Force sync status change for monitored objects
For optimization purposes, we don't fire DB calls to update a status
that is not changing compared to the previous version. For managed
objects transitioning to monitored however this could be critical,
as the Hashtree builder expects a status update in order to remove
the nodes from all the trees.