Manage object status asynchronously
As AIM becomes a schema instead of a library, let AID manage
completely status objects and their state, so that we don't
need to rely on northbound orchestration systems to do so.

This change reaches this goal with the following:

- AIM now doesn't set any Status object for newly created/updated
  objects;
- Status default sync_state is now N/A (covers aimctl get calls);
- On hashtree differences, AID sets the SYNC_PENDING state;
- SYNC_PENDING objects are represented by a new metadata called
  "pending" in the hashtree;
- Hashtree nodes that are converged, but with the "pending"
  attribute set, are put in SYNCED state by AID;
- SYNC_ERROR continues to work as before.

NOTE: Monitor state update can now be unreliable when multiple AIDs
are running. Will be addressed by future commits