Reset k8s watcher retries number after a clean monitoring iteration
when observe_and_monitor_loop runs, it only exits the function when
an error occurs or the thread is stopped. This because it's blocked
in a loop internal to the one already provided by the decorator.
Because of that, whenever an error occurs its retry count is increased
by the decorator without ever having a chance to be reset. This
change makes sure that the only loop in this method is the
decorator itself.

In addition, save hashtrees from the K8S watcher every time for
operational resources, even if they look like empty update, so
that we don't lose information about the node error state.