Include endpoints in injected Service objects
Most of the change is mechanical except for
handling of service endpoints with Kubernetes
store since Kubernetes saves endpoints in a
different object ("Endpoints") instead of the
"Service" object. Currently we support
read-only access to service endpoints with
Kubernetes store. Also kubernetes watcher is
modified to listen for Endpoints events and
modify them into Service events.

Signed-off-by: Amit Bose <amitbose@gmail.com>