Don't put objects in error state while changing monitored value
Whenever an AIM object switches its monitored state, the object
is moved from the Monitored tree (current state) to the Config tree
(desired state). Doing this causes the following:

1) Configuration universe created the TAG on ACI for the object;
2) Monitored universe tries to re-create the object in AIM with
   monitored = True: this fails by design;
3) Events arrive for #1, and on the ACI universe, the object is
   now in the Config tree (current state) so on the next AID
   iteration #2 doesn't happen anymore. Everything is synced.

If #2 happens before #1 however, then the object is put in Error
state, preventing #2 from happening (the AIM monitored state of the
object is still False. So the affected tree will be the Config
even though this happened in the Monitored Universe!). This causes
divergence until the divergence control engine kicks in and surrenders
the operation.

To avoid the problem, we shouldn't put the object in error state
during #2.