Implement cascade object deletion
Add cascade key argument to the AIM manager delete method. This
will delete an object's subtree entirely. Can be used to cleanup
a specific ACI root.

For consistency, we don't return errors when delete is requested for a
non existing object, but we delete its subtree regardless if any resource
exist in it.

Closes noironetworks/support #590