Add support for multiple VMM types
Two new attributes are added to the EPG resource:

- vmm_domains: list of dictionaries describing VMMs to which the
EPG refers to;
- physical_domains: list of dictionaries describing physdoms to
which the EPG refers to.

right now, only name and type (in the case of VMM) can be specified.
In future, we can add extra attributes to custom VMM capabilities.
the old attributes 'physical_domain_names' and 'openstack_vmm_names'
are still around and can be used both for reading and writing
new references, but their value will be mirrored in the vmm_domains
and physical_domains attributes.