Add the pod_id and from_config column to HostLinks table
We need the from_config column to indicate if this table entry was created
from the config file or not. We need this because in apicapi there is a
clear_staticlinks() API which will clean up those entries whose interface_name
field starts with 'static'. This will not work well once we allow user to put the
valid interface_name into the config file so we will have to use another field to
do the clear_staticlinks() job.