Generalize HostDomainMapping table
The vmm_domain_name column is changed to just domain_name and made
a primary key, while the physical_domain_name is dropped in favor
of a domain_type, which is an enum for the supported Domain types.
This is needed in order to allow multiple entries with the same
hostname in the table, which can be used to support a catch-all
or "wildcard" of all hosts for multiple VMM domains.