Properly skipp root deletion for monitored objects
Today, the monitored universe skips the deletion of root objects
with a warning message that can often be confused with actual divergence.
On top of that, the action cache logs all these deletion attempts as
if they were part of an unfixable divergence. We should skip the
deletion at a previous time in the reconciliation cycle, so that
AID still doesn't "smile" (after all, we are monitoring a
non-existent object) but so that we don't cause divergence either.