NAT library: Don't take ownership of L3Outs
In the no NAT strategy, we would take ownership of
L3Outs with monitored=True so that we can set the
VRF associated with it. This causes issues with
sharing of L3Outs between different AIM/AID systems,
and is probably best avoided. Instead, an exception
is now raised if an attempt is made to connect
the L3Out to a different VRF.

The DB migration code also relinquishes ownership of
L3Outs that we had owned, and deletes the owenrship
tag from APIC. The latter is done so that we can
completely remove support for monitored state changes
(which has been a source of bugs).

Signed-off-by: Amit Bose <amitbose@gmail.com>