Change tenant removal algorithm.
The previous algorithm for tenant removal was proved to be unreliable
in situations of high load, where there are lots of threads and race
conditions are likely. This would often cause leftover objects laying
around, since the monitored universe would keep syncing in the background
without causing the removal to cease.

With the new algorithm, we have a two phase decision that is taken
before and after the reconciliation cycle. Furthermore, we make sure
that the monitored universes delete all the AIM monitored objects
when they agree on tenant removal. We do so by pretending that the
desired state is empty. Only when a tenant's AIM state is completely
empty (all trees with no root) we finally cleanup the state.
DB transactionality and update lock should ensure this happens
reliably.