Add VMM domain configuration to L3 out
The existing API for L3 Out configuration doesn't allow the
caller to specify the VMM or Physical Domains that should be
associated to the external EPG created for the L3 Out. The
EPG is then associated with all VMM domains. This commit
adds the ability to control VMM and Physical Domain asociation.