Add load-mappings command
The load-mappings command is added, so that the information
specified in the configuration file can be used to create
entries in the HostDomainMappingV2 table. This patch also
adds a call to this command in the load-domains command,
so that the mappings can be created when the domains first
get loaded.