Dealing with ResourceClosedError
An AIM DB table, ActionLog, is populated in the neutron
server thread, while an AID thread consumes from this table
and deletes rows. During the DB object versioning change in AIM,
an “after_transaction_end” hook was added which, among other
things, processes ActionLog DB objects that were changed during
that transaction. If AID runs and deletes a row corresponding to an
ActionLog DB that is subsequently processed in the
after_transaction_end hook, it raises an ObjectDeletedException, and
sqlalchemy rolls back. We actually don’t see this exception since
sqlalchemy eats it up.

This patch proposed to ignore the ObjectDeletedException and
allow the post-commit processing to proceed.