[WIP}: Decouple SecuritGroupRuleRemoteIp in ORM
The SecurityGroupRule defines a list of SecurityGroupRuleRemoteIp
entries in the ORM, using a relationship. List updates are costly
at the DB level, and therefore should be decoupled. The relationship
is removed, and the SecurityGroupRuleRemoteIp table is exposed as
a resource. A DB migration is also created to add the epoch column
to the SecurityGroupRuleRemoteIp entries, in order to help with
concurrency. This change introduces backwards-incompatible semantics,
as the remote_ips attribute of the SecurityGroupRule is no longer
a reflection of the associated rows in the SecurityGroupRemoteIp
table.