Fix StaleDataError exceptions
This patch relaxes serializtion of changes to
SecurityGroupRuleRemoteIp table entries by removing updates of the
epoch in SecurityGroupRule entries when there are no changes to the
other rows of the SecurityGroupRule itself. A corresponding fix is
needed for the processing of the associated ActionLog entries, in order
to ensure that correctness is maintained in the face of concurrent
updates to the same SecurityGroupRule entry.