Support indexed utf-8 encoded columns
Some database backends are constrained to a size of 767 bytes for
indexed columns. Many backends use utf-8 for character encoding, which
can use a maximum of 3 bytes per character. This triggers warnings
or errors when running DB migrations. This patch reduces these columns
to a size of 255 bytes. Note that if utf8mb4 character encoding is
used, then these as well as other columns must be reduced to 191 bytes
(and would be fixed by a separate patch), as it uses 4 bytes per character