Make sure the APIC rotation will work
1. Have to reset all the tenants once a rotation happens. This
will make those tenants re-subscribe again with the new APIC
connection.
2. No need to subscribe again for each tenant when there is an
APIC event available. This is problematic because if this
subcribe fails, each tenant will try to re-establish a new APIC
connection but then other tenants are not getting notified to
re-subscribe. There is only one APIC connection for all the tenants
so this should not be handled by each tenant.