Distribute APICs evenly among those aim-aid controllers
1. Use a new DB table to coordinate the APIC assignments
to make sure they are distributed evenly.

2. When the assigned APIC goes down, aim-aid will try
to find the next available APIC or fall back to one
APIC which has already been assigned to somebody else.
For latter case, it will then enter the recovery mode
and try to find the next available APIC periodically
with a back-off timer.