Override default for use_stderr to False
In older releases of oslo_log, the use_stderr cfg_opt has a default
of True. This results in logs getting replicated to /var/log/messages
in addition to the targetted log file as seen in newton deployments.
Being unnecessary duplication the community has fixed this in later
releases, we set it to False in the aim.conf template for installers
to pick up.