Add mode for EPG static ports
The current code for supporting static port configuration on EPGs
lacks the ability to configure the mode. ACI offers 3 modes:

1) "regular". This is a trunked configuration
2) "native". This is an 802.1p configuration
3) "untagged". This is an untagged configuration with a native VLAN

This patch adds the ability to configure the mode setting, and uses
a default setting of "regular".  If the mode setting isn't provided,
then the static port is configured using "regular", but is not reflected
in the static_paths list object, in order to maintain backwards-compatibility.
Non-default settings ("native" or "untagged") will be reflected in the
list objects.

The patch includes a DB migration, which adds this column to the table,
and populates rows in the table with the default value ("regular"), since
this was the default value for any such configuration in AIM.