Add the remote_group_id column to the SG_rule table
1. Also prevent the SG rule MO from showing up in APIC by
removing it from the AIM tree when it has remote_group_id
set and its remote_ips list is empty. Also add a temporary
config knob to control this behavior in case user would
like to roll back to the previous behavior somehow.

2. Also remove the temporary config parameter fetch_sgr_from_db
because we have been using it for a while without issues + we
will need to do that always now for item #1 to work.

3. Also return the hash tree result properly in the
_get_hash_trees() API otherwise the test_tree_result() is a
no-op.