Remove use of with_lockmode
The use of the now-removed with_lockmode parameter is meaningless,
since most deployments use a distributed database (e.g. Galera)
that doesn't support distributed locking.

Commit 462b1ea711c91b06bd5e8c5cfcd3e8c3a1d480e5 pinned SQLAlchemy
in order to avoid the exception that was thrown due to the fact that
the "with_lockmode" parameter was removed from newer versions.

This patch un-pins SQLAlchemy and removes the use of with_lockmode.