Fix websocket subscription refreshes
AIM uses the websocket refresh subscription mechanism implemented
in acitoolkit. However, it uses a hard-coded refresh interval (30
seconds), with all subscriptions refreshed at the same time. This
results in bursts of subscription refreshes, and causes excessive
load on the APIC controller when there is a large number of
roots/tenants. The subscription refresh API in acitoolkit was
extended to support passing in URLs for refresh, which allows each
AciTenantManager object to handle their own refreshes.

This patch also ensures that the refresh interval isn't set to
lower than the default for APIC (60 seconds).