Add APIC login class
The apicapi client requires logging in to APIC under certain
conditions, such as continued authorization failures. This patch
adds a class that attempts to reauthorize/login with APIC if the
auth token becomes invalidated. The authorization class tries to
refresh the login within half of the login timeout period that
was returned by APIC in the authorization response/token.