Fix use of ExternalNetwork contracts
The following commits made contract references under the
ExternalNetwork resource top-level resources:
*  8a8c6911a073d20d25858e39e6ab66ce759a5ea9
*  747ab9b43db4f978a7f146e0279b7e52da3cb480
*  76cfa143d3c2be8843e949e813361a3c84651142
*  adf31160ca50efdeba40b42a667cfed225828dc7

However, it preserved the semantics of resource creation and deletion
using the provided_contracts and consumed_contracts members of the
ExternalNetwork, which made it possible to hit the same issue outlined
in 8a8c6911a073d20d25858e39e6ab66ce759a5ea9. This patch removes the
semantics of contract creation using the lists in the ExternalNetwork
resource, which means AIM applications must be changed to use the
top-level resources.