Removing persistance of hashtrees in DB
This removes persisting the hashtrees in DB. Instead, the hashtrees are saved
in-memory. We do still use the DB for the hash tree roots. This is done to make
the impact minimal and not to affect the current workflow of AIM. This helps
reduce the  size of hash tree BLOBs in the DB, as they were getting too big. It
also reduces the DB load.