Add GARP configuration for gen1 hardware
First generation switches are unable to use GARPs for EP movement
detection in hardware when the EP "move" is on the same port..As a
result, additional configuraiton is needed in the BD to punt all GARPs
to the switch CPU for processing. This patch adds an AIM parameter
to control how this behavior is configured in the fabric, with the
default configuration to disable it (i.e. default is to not support
use of GARPs for EP movement detection on gen1 hardware).