Fix stale tenant cleanup
Commit 4fb4cabf8a7e0cbb9546b892d4a3867a19fa0d04 was created to avoid
cases where stale tenant state is left in APIC, due to hash trees being
deleted before the reconcile could create the delete requests to APIC.
The fix failed to cover the case where tenants are created and then
deleted before a reconcile happens, and would fail to delete the hash
trees and root, causing AIM to consider the stale trees during the
reconciliation cycle. This patch uses an approach to wait 3
cycles per universe (config, operational, and monitored) before
allowing tenant deletion to happen for that universe.  Note that
all 3 universes must allow the tenant state to be deleted, making
for up to 9 reconciliation cycles before a tenant is finally
removed.