Fix DB migration
Commit 94ab02aadf7986023b70074db4568e115b4ab5b4 (and the fix in
cb8847a610a878fe22d823ad537344f615c3f503) added a database migration
for a new feature. This migration fails in older versions of the
DB layer, due to an invalid primary key length. As a result, the existing
migration has to be fixed to use a supported column size.

Although the broken migration was released, the feature was never
socialized and therefore never used, which means that the table with
the invalid primary key length is empty. An additional DB migration
is created that drops the original table, and re-creates it with the
valid key length. This ensures that deployments using the older DB
layer software can be upgraded to newer software, and both will have
the same key length.